+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## 1.0.2
+
+- Renamed metrics to logger.
+- Added `removeIndexes` on `RxList`.
+- Improved `RxList` remove items performance.
+
 ## 1.0.1
+
 - Fixed debug name parsing for minified code
 - Added option to disable global logging
 - Disabled global logging by default
 
-## 1.0.0. 
+## 1.0.0
+
 - Initial release.